modifique un error en el uml,edite el tipo de algunas caracteristicas de los objetos y añadi el menu de busquedas,solucion error de inicio sesion

+++ src/main/java/Entrega.java
@@ -1,12 +1,12 @@
 public class Entrega {
     public int Cod_Entrega;
-    public int Peso;
+    public double Peso;
     public int Dimenciones;
     public boolean Fragil;
     public double Costo;
     public String Estado_Entrega;
 
-    public Entrega(int cod_Entrega, int peso, int dimenciones, boolean fragil, double costo, String estado_Entrega) {
+    public Entrega(int cod_Entrega, double peso, int dimenciones, boolean fragil, double costo, String estado_Entrega) {
         Cod_Entrega = cod_Entrega;
         Peso = peso;
         Dimenciones = dimenciones;
