+++ src/main/java/Repartidor.java
@@ -2,11 +2,11 @@ public class Repartidor {
     public int Cedula;
     public String Nombre;
     public int Edad;
-    public double Num_Celular;
+    public int Num_Celular;
     public String Tipo_Licencia;
     public double Sueldo;
 
-    public Repartidor(int cedula, String nombre, int edad, double num_Celular, String tipo_Licencia, double sueldo) {
+    public Repartidor(int cedula, String nombre, int edad, int num_Celular, String tipo_Licencia, double sueldo) {
         Cedula = cedula;
         Nombre = nombre;
         Edad = edad;
